Work with Outlook 15.38
Without this fix, when I run the workflow, I get nothing in the Alfred
workflow debug pane, but in Console.app, I see:

```
default	09:42:24.815664 -0700	Microsoft Outlook	An exception was thrown during execution of an NSScriptCommand...
default	09:42:24.815747 -0700	Microsoft Outlook	[<ECMessage 0x600002a26040> valueForUndefinedKey:]: this class is not key value coding-compliant for the key edited.
```

So this looks like perhaps a bug in Outlook that it is not key value
coding-compliant, but this simple change seems to make it work.